# public/openapi.v1.draft.yaml
openapi: 3.1.0
info:
  title: School Timetable & Attendance API
  version: 1.0.0-draft
  contact:
    name: Szkola API Maintainers
    url: https://szkola.tkch.eu
  description: |
    Draft of a developer-friendly API for timetable browsing, attendance tracking,
    remote approvals, and admin overrides. Uses resource-oriented endpoints,
    standard HTTP semantics, and consistent error handling.
servers:
  - url: https://api.example.com
  - url: http://localhost:8787
...
security:
  - BearerAuth: []
  - CookieAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API-Key
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth
  parameters:
...
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: false
      schema: { type: string }
      description: Required when using cookie authentication. Must match csrf cookie value.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema: { type: string }
      description: Optional key to safely retry a non-idempotent request.
  headers:
...
    RateLimitLimit:
      description: Maximum number of requests allowed in the current window.
      schema: { type: integer }
    RateLimitRemaining:
      description: Remaining requests in the current window.
      schema: { type: integer }
    RateLimitReset:
      description: Seconds until the rate-limit window resets.
      schema: { type: integer }
  responses:
    Problem400:
      description: Bad request
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Problem401:
      description: Unauthorized
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Problem403:
      description: Forbidden
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Problem404:
      description: Not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Problem409:
      description: Conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Problem410:
      description: Gone
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Problem429:
      description: Too many requests
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Problem500:
      description: Internal server error
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: RFC 7807-like problem details.
      properties:
        type: { type: string, description: Unique error identifier URI or code }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code: { type: string, description: Domain-specific, machine-readable code }
        fields:
          type: array
          description: Field-level validation errors
          items:
            type: object
            properties:
              path: { type: string }
              message: { type: string }
...
    ApprovalToken:
      type: object
      properties:
        token: { type: string }
        url: { type: string }
        expiresAt: { type: string, format: date-time }
      required: [token, url, expiresAt]
...
    JobStatus:
      type: object
      properties:
        id: { type: string }
        kind: { type: string, nullable: true }
        status: { type: string, enum: [queued, running, succeeded, failed, timeout] }
        createdAt: { type: string, format: date-time }
        startedAt: { type: string, format: date-time, nullable: true }
        finishedAt: { type: string, format: date-time, nullable: true }
        error: { type: string, nullable: true }
      required: [id, status]
...
paths:
...
  /v1/login:
    post:
      security: []
      summary: Log in and set a session cookie (for web panel)
      operationId: loginUser
      tags: [Auth]
      description: Authenticates a user and issues the auth session cookie.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
              required: [username, password]
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      authenticated: { type: boolean }
                    required: [authenticated]
                required: [ok, data]
        '401':
          $ref: '#/components/responses/Problem401'
        '429':
          $ref: '#/components/responses/Problem429'
...
  /v1/register:
    post:
      security: []
      summary: Register a new user and create a session (when registration is enabled)
      operationId: registerUser
      tags: [Auth]
      description: Creates a new user account (when enabled) and starts an authenticated session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                password: { type: string }
              required: [username, password]
      responses:
        '200':
          description: Registered
          headers:
            RateLimit-Limit: { $ref: '#/components/headers/RateLimitLimit' }
            RateLimit-Remaining: { $ref: '#/components/headers/RateLimitRemaining' }
            RateLimit-Reset: { $ref: '#/components/headers/RateLimitReset' }
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      authenticated: { type: boolean }
                    required: [authenticated]
                required: [ok, data]
        '400':
          description: Invalid username or password
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '403':
          description: Registration disabled
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409':
          description: User exists
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '429':
          description: Too many registration attempts
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '500':
          $ref: '#/components/responses/Problem500'
...
  /v1/approvals:
    post:
      security:
        - CookieAuth: []
      summary: Create a one-time approval token to update an entry
      operationId: createApprovalToken
      tags: [Approvals]
      description: Creates a one-time approval token for remote accept or deny flow.
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                action:
                  type: string
                  enum: [toggle, set]
                dateISO:
                  type: string
                  format: date
                entryId: { type: string }
                present:
                  type: boolean
                  description: Required when action = set
              required: [action, dateISO, entryId]
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data: { $ref: '#/components/schemas/ApprovalToken' }
                required: [ok, data]
        '400':
          $ref: '#/components/responses/Problem400'
        '401':
          $ref: '#/components/responses/Problem401'
        '403':
          $ref: '#/components/responses/Problem403'
        '500':
          $ref: '#/components/responses/Problem500'
  /v1/approvals/{token}:
    get:
      security: []
      summary: Get approval status
      operationId: getApprovalStatus
      tags: [Approvals]
      description: Returns current status and expiry metadata for an approval token.
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Approval status
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      status: { type: string, enum: [pending, accepted, denied, expired] }
                      createdAt: { type: string, format: date-time }
                      expiresAt: { type: string, format: date-time }
                    required: [status]
                required: [ok, data]
        '404':
          $ref: '#/components/responses/Problem404'
        '410':
          $ref: '#/components/responses/Problem410'
        '500':
          $ref: '#/components/responses/Problem500'
    post:
      security: []
      summary: Submit a decision for an approval token
      operationId: decideApproval
      tags: [Approvals]
      description: Applies accept or deny decision to a pending approval token.
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                decision: { type: string, enum: [accept, deny] }
              required: [decision]
      responses:
        '200':
          description: Decision accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      status: { type: string, enum: [accepted, denied] }
                    required: [status]
                required: [ok, data]
        '400':
          $ref: '#/components/responses/Problem400'
        '403':
          $ref: '#/components/responses/Problem403'
        '404':
          $ref: '#/components/responses/Problem404'
        '409':
          $ref: '#/components/responses/Problem409'
        '410':
          $ref: '#/components/responses/Problem410'
        '500':
          $ref: '#/components/responses/Problem500'
...
  /v1/jobs/timetable-scrape:
    post:
      security:
        - CookieAuth: []
      summary: Start an async job to refresh timetable data (admin-only)
      operationId: startTimetableScrapeJob
      tags: [Jobs]
      description: Starts or reuses an asynchronous timetable scraping job (admin).
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '202':
          description: Job accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      jobId: { type: string }
                      statusUrl: { type: string }
                      status: { type: string, enum: [queued, running, succeeded, failed, timeout] }
                    required: [jobId, statusUrl, status]
                required: [ok, data]
        '403':
          description: Admin required
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '401':
          $ref: '#/components/responses/Problem401'
        '500':
          $ref: '#/components/responses/Problem500'
  /v1/jobs/articles-scrape:
    post:
      security:
        - CookieAuth: []
      summary: Start an async job to refresh articles (admin-only)
      operationId: startArticlesScrapeJob
      tags: [Jobs]
      description: Starts or reuses an asynchronous article scraping job (admin).
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '202':
          description: Job accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data:
                    type: object
                    properties:
                      jobId: { type: string }
                      statusUrl: { type: string }
                      status: { type: string, enum: [queued, running, succeeded, failed, timeout] }
                    required: [jobId, statusUrl, status]
                required: [ok, data]
        '403':
          description: Admin required
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '401':
          $ref: '#/components/responses/Problem401'
        '500':
          $ref: '#/components/responses/Problem500'
  /v1/jobs/{jobId}:
    get:
      security: []
      summary: Get job status
      operationId: getJobStatus
      tags: [Jobs]
      description: Returns current state and metadata for an asynchronous job.
      parameters:
        - name: jobId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  data: { $ref: '#/components/schemas/JobStatus' }
                required: [ok, data]
        '404':
          $ref: '#/components/responses/Problem404'
...